Separate Zlib header validation fom stream reading
